ICU-12701 more IBM i test fixes.
* intltest now builds.
* fix for unpax-icu.sh to include the utf-16be file.
* added an emacs coding line to 'encoded.utf16be'

X-SVN-Rev: 39304

+++ icu4c/source/test/intltest/itrbnf.cpp
@@ -2124,7 +2124,7 @@ void IntlTestRBNF::TestPluralRules() {
 
     // Make sure there are no divide by 0 errors.
     UnicodeString result;
-    RuleBasedNumberFormat(ruRules, Locale("ru"), parseError, status).format(21000, result);
+    RuleBasedNumberFormat(ruRules, Locale("ru"), parseError, status).format((int32_t)21000, result);
     if (result.compare(UNICODE_STRING_SIMPLE("twenty-one thousand")) != 0) {
         errln("Got " + result + " for 21000");
     }
