+++ icu4c/source/test/intltest/pluralmaptest.cpp
@@ -56,24 +56,24 @@ void PluralMapTest::runIndexedTest(int32_t index, UBool exec, const char* &name,
 }
 
 void PluralMapTest::TestToCategory() {
-    assertEquals("", PluralMapBase::OTHER, PluralMapBase::toCategory("other"));
-    assertEquals("", PluralMapBase::ZERO, PluralMapBase::toCategory("zero"));
-    assertEquals("", PluralMapBase::ONE, PluralMapBase::toCategory("one"));
-    assertEquals("", PluralMapBase::TWO, PluralMapBase::toCategory("two"));
-    assertEquals("", PluralMapBase::FEW, PluralMapBase::toCategory("few"));
-    assertEquals("", PluralMapBase::MANY, PluralMapBase::toCategory("many"));
-    assertEquals("", PluralMapBase::NONE, PluralMapBase::toCategory("Many"));
+    assertEquals("", (int32_t)PluralMapBase::OTHER, PluralMapBase::toCategory("other"));
+    assertEquals("", (int32_t)PluralMapBase::ZERO, PluralMapBase::toCategory("zero"));
+    assertEquals("", (int32_t)PluralMapBase::ONE, PluralMapBase::toCategory("one"));
+    assertEquals("", (int32_t)PluralMapBase::TWO, PluralMapBase::toCategory("two"));
+    assertEquals("", (int32_t)PluralMapBase::FEW, PluralMapBase::toCategory("few"));
+    assertEquals("", (int32_t)PluralMapBase::MANY, PluralMapBase::toCategory("many"));
+    assertEquals("", (int32_t)PluralMapBase::NONE, PluralMapBase::toCategory("Many"));
     assertEquals(
             "",
-            PluralMapBase::FEW,
+            (int32_t)PluralMapBase::FEW,
             PluralMapBase::toCategory(UnicodeString("few")));
     assertEquals(
             "",
-            PluralMapBase::MANY,
+            (int32_t)PluralMapBase::MANY,
             PluralMapBase::toCategory(UnicodeString("many")));
     assertEquals(
             "",
-            PluralMapBase::NONE,
+            (int32_t)PluralMapBase::NONE,
             PluralMapBase::toCategory(UnicodeString("Many")));
 }
 
@@ -117,24 +117,24 @@ void PluralMapTest::TestIterate() {
     PluralMapBase::Category index = PluralMapBase::NONE;
     const UnicodeString *current = map.next(index);
     assertEquals("", "pickles", *current);
-    assertEquals("", PluralMapBase::OTHER, index);
+    assertEquals("", (int32_t)PluralMapBase::OTHER, index);
     current = map.next(index);
     assertEquals("", "pickle", *current);
-    assertEquals("", PluralMapBase::ONE, index);
+    assertEquals("", (int32_t)PluralMapBase::ONE, index);
     current = map.next(index);
     assertEquals("", "picklefew", *current);
-    assertEquals("", PluralMapBase::FEW, index);
+    assertEquals("", (int32_t)PluralMapBase::FEW, index);
     current = map.next(index);
-    assertEquals("", PluralMapBase::CATEGORY_COUNT, index);
+    assertEquals("", (int32_t)PluralMapBase::CATEGORY_COUNT, index);
     assertTrue("", current == NULL);
 
     PluralMapForPluralMapTest map2;
     index = PluralMapBase::NONE;
     current = map2.next(index);
     assertEquals("", "", *current);
-    assertEquals("", PluralMapBase::OTHER, index);
+    assertEquals("", (int32_t)PluralMapBase::OTHER, index);
     current = map2.next(index);
-    assertEquals("", PluralMapBase::CATEGORY_COUNT, index);
+    assertEquals("", (int32_t)PluralMapBase::CATEGORY_COUNT, index);
     assertTrue("", current == NULL);
 }
